xenial: drop -updates because they bring in pain.
Ubuntu 16.04 is not exactly the same thing as it was on the moment of
launch.  The so-called HWE updates bring in tech stacks from newer
releases - such as new kernel, or new xorg/mesa stack.  This also means
that the bugs introduced in newer distributions are now seen in 16.04 as
well - such as the infamous protobuf library bug where it can not be
unloaded if loaded twice.

To fix that, we need to make sure we're using only security-updated
Ubuntu, and not to pull in any dependencies that depend on mir.