Add new image for libplacebo based on ubuntu cosmic
We use ubuntu rather than debian because the former has a package for
the vulkan validation layers whereas the latter appears not to.
Unfortunately, even they don't support libshaderc, but we've got to work
with what we've got - so we install it by hand. (Taking care not to shit
all over the glslang package's contents)

Incidentally, while I install the mesa vulkan drivers, they probably
won't actually work in the test suite unless we actually have any GPU
inside the VM. But that's a separate concern.