libplacebo-ubuntu-cosmic: fix installation of shaderc .so
libshaderc_shared.so is a symlink to libshaderc_shared.so.1. `install`
aggressively dereferences symlinks, so it ended up installing the .so.1
to the path of .so, which worked while resolving dependencies but broke
at link time - resulting in a failed job.

Fix this issue by just using `cp -a` instead - the permissions are
correct by default anyway.