kaniko: redo the image in a proper way.
Previously, we were using stretch base image which mostly worked, but in
some cases the files werent propagated to the destination image.  The
proper way is to have a really minimal kaniko image, with
/kaniko/executor and a shell.  We could really use just that, but to
figure out which images we should build, we also require git - and
that's the tricky part.
So, we start with alpine image, copy all files into /alpine/ and fix
busybox to provide shell.