medialibrary-win: Bump meson to a more recent version
For win64 this will allow us to generate a json test summary.
For all platforms, it also has a better display of the tests results
when run in parallel, while the version we're using on win32 is dumping
all the results at the end.
I'm also hoping that this updated version will fix some of the sporadic
failures.