libplacebo-ubuntu-jammy: install lld alongside clang
This is not automatically pulled in because it's not listed as a hard
dependency, just a recommendation.

Needed for https://code.videolan.org/videolan/libplacebo/-/merge_requests/429

As an aside, drop the specific version from the jammy tag used as base -
this is pointless because we do an `apt upgrade` in the docker build
process anyway.